rosrust

Struct Duration

pub struct Duration {
    pub sec: i32,
    pub nsec: i32,
}

ROS representation of duration, with nanosecond precision

Fields
sec: i32

Number of seconds. Negative for negative durations.

nsec: i32

Number of nanoseconds inside the current second. Negative for negative durations.

Implementations

impl Duration

pub fn new() -> Duration

Creates a new duration of zero value.

Examples

assert_eq!(Duration::new(), Duration { sec: 0, nsec: 0 });

pub fn from_nanos(t: i64) -> Duration

Creates a duration of the given number of nanoseconds.

Examples

assert_eq!(Duration::from_nanos(0), Duration { sec: 0, nsec: 0 });
assert_eq!(Duration::from_nanos(12_000_000_123), Duration { sec: 12, nsec: 123 });
assert_eq!(Duration::from_nanos(-12_000_000_123), Duration { sec: -12, nsec: -123 });

pub fn from_seconds(sec: i32) -> Duration

Creates a duration of the given number of seconds.

Examples

assert_eq!(Duration::from_seconds(0), Duration { sec: 0, nsec: 0 });
assert_eq!(Duration::from_seconds(12), Duration { sec: 12, nsec: 0 });
assert_eq!(Duration::from_seconds(-12), Duration { sec: -12, nsec: 0 });

pub fn nanos(self) -> i64

Returns the number of nanoseconds in the duration.

Examples

assert_eq!(Duration { sec: 0, nsec: 0 }.nanos(), 0);
assert_eq!(Duration { sec: 12, nsec: 123 }.nanos(), 12_000_000_123);
assert_eq!(Duration { sec: -12, nsec: -123 }.nanos(), -12_000_000_123);

pub fn seconds(self) -> f64

Returns the number of seconds in the duration.

Examples

assert_eq!(Duration { sec: 0, nsec: 0 }.seconds(), 0.0);
assert_eq!(Duration { sec: 12, nsec: 123 }.seconds(), 12.000_000_123);
assert_eq!(Duration { sec: -12, nsec: -123 }.seconds(), -12.000_000_123);