rosrust::api::handlers

Trait SubscriptionHandler

pub trait SubscriptionHandler<T>: Send + 'static {
    // Required methods
    fn connection(&mut self, headers: HashMap<String, String>);
    fn message(&mut self, message: T, callerid: &str);
}

Handles all calls involved with a subscription

Each subscription is done in one thread, so there is no synchronization necessary between these calls.

Required Methods

fn connection(&mut self, headers: HashMap<String, String>)

Called before any message is accepted from a certain caller ID

Contains the headers for handling the specific connection

fn message(&mut self, message: T, callerid: &str)

Called upon receiving any message